requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-background-tasks==1.2.5
django-compat==1.0.15
djangorestframework==3.11.1
httpie==2.2.0
idna==2.10
paho-mqtt==1.5.0
Pygments==2.6.1
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
